requirements.txt
hikari==2.0.0.dev105
hikari-tanjun==2.3.0a1
prisma==0.4.3
pydantic==1.9.0
